amended setup script and check endpoint script

+++ setup_script.py
@@ -8,57 +8,10 @@ Description   : This script will create cluster prefix folders and copy relevant
 """
 
 import argparse
-import logging
 import os
 import shutil
 import sys
-import types
 from datetime import datetime
-
-
-def log_newline(self, how_many_lines=1):
-
-    # Switch formatter, output a blank line
-    self.handler.setFormatter(self.blank_formatter)
-
-    for i in range(how_many_lines):
-        self.info("")
-
-    # Switch back
-    self.handler.setFormatter(self.formatter)
-
-
-def create_logger(cluster_prefix):
-
-    # Create a handler
-    sh = logging.StreamHandler(sys.stdout)
-    handler = logging.FileHandler(
-        f"{datetime.now().strftime('%Y-%m-%d_%H-%M-%S')}_{cluster_prefix}-setup_script.log",
-        mode="w",
-        encoding="utf-8",
-    )
-    handler.setLevel(logging.DEBUG)
-    formatter = logging.Formatter(
-        fmt="[%(asctime)s] %(levelname)8s : %(message)s", datefmt="%Y-%m-%d %H:%M:%S"
-    )
-    blank_formatter = logging.Formatter(fmt="")
-    handler.setFormatter(formatter)
-
-    # Create a logger, with the previously-defined handler
-    logger = logging.getLogger("logging_test")
-    logger.setLevel(logging.DEBUG)
-    logger.addHandler(handler)
-    sh.setFormatter(formatter)
-    logger.addHandler(sh)
-
-    # Save some data and add a method to logger object
-
-    logger.handler = handler
-    logger.formatter = formatter
-    logger.blank_formatter = blank_formatter
-    logger.newline = types.MethodType(log_newline, logger)
-
-    return logger
 
 
 def copy_scripts(cluster_prefix):
@@ -76,8 +29,14 @@ def copy_scripts(cluster_prefix):
     ]
 
     for script in script_list:
-        logger.info(f"Copying script {script}....")
+        script = os.path.join("./scripts", script)
+        if not os.path.exists(script):
+            print(f"Failed: Script {script} does not exist.")
+            sys.exit(1)  # Exit with error status
+        print(f"Copying script {script}....")
         shutil.copy(script, cluster_prefix)
+    
+    return True
 
 
 def create_directory(cluster_prefix):
@@ -91,15 +50,20 @@ def create_directory(cluster_prefix):
         try:
             os.mkdir(dirpath)
         except FileExistsError:
-            logger.warning(f"Directory {dirpath} already exists")
+            print(f"Directory {dirpath} already exists")
+        except Exception as e:
+            print(f"Failed to create directory {dirpath}: {e}")
         else:
-            logger.info(f"Directory {dirpath} created")
+            print(f"Directory {dirpath} created")
+
+    return True
 
 
 def main():
 
-    create_directory(cluster_prefix)
-    copy_scripts(cluster_prefix)
+    if create_directory(cluster_prefix):
+        if copy_scripts(cluster_prefix):
+            print("Setup completed successfully.")
 
 
 if __name__ == "__main__":
@@ -112,14 +76,11 @@ if __name__ == "__main__":
         help="Specify the cluster prefix for the folder name. Example rk-dt or pb-gen-preprod.",
     )
     if len(sys.argv) != 3:
-        print("USAGE: python setup_script.py --c <cluster_prefix>")
         print("ERROR: Please provide the relevant input to run.")
+        print("USAGE: python setup_script.py --c <cluster_prefix>")
         sys.exit(1)  # Exit with error status
 
     args = parser.parse_args()
     cluster_prefix = args.cluster_prefix
-
-    # Set global logger
-    logger = create_logger(cluster_prefix)
 
     main()
